:hammer: relocate json_dumps
