+++ pyecharts/javascript.py
@@ -1,3 +1,7 @@
+# coding=utf-8
+import json
+import datetime
+
 import pyecharts.constants as constants
 import pyecharts.exceptions as exceptions
 
@@ -17,6 +21,7 @@ def clear():
 def add_a_new_function(afunc):
     if not constants.PY35_ABOVE:
         raise exceptions.JavascriptNotSupported(constants.ERROR_MESSAGE)
+
     CUSTOM_FUNCTIONS[afunc.__name__] = afunc
     return FUNCTION_SIGNATURE % afunc.__name__
 
@@ -32,7 +37,7 @@ def has_functions():
     return len(CUSTOM_FUNCTIONS) > 0
 
 
-def compile():
+def translate_python_functions():
     try:
         from pyecharts_javascripthon import Python2Javascript
     except ImportError:
@@ -44,3 +49,37 @@ def compile():
             CUSTOM_FUNCTIONS[func])
         content.append(javascript_function)
     return ''.join(content)
+
+
+class UnknownTypeEncoder(json.JSONEncoder):
+    """
+    UnknownTypeEncoder`类用于处理数据的编码，使其能够被正常的序列化
+    """
+
+    def default(self, obj):
+        if isinstance(obj, (datetime.datetime, datetime.date)):
+            return obj.isoformat()
+        else:
+            # Pandas and Numpy lists
+            try:
+                return obj.astype(float).tolist()
+            except Exception:
+                try:
+                    return obj.astype(str).tolist()
+                except Exception:
+                    return json.JSONEncoder.default(self, obj)
+
+
+def translate_options(data, indent=0):
+    """ json 序列化编码处理
+
+    :param data: 字典数据
+    :param indent: 缩进量
+    """
+    options_in_json = json.dumps(data, indent=indent, cls=UnknownTypeEncoder)
+    if has_functions():
+        options_with_js_functions = unescape_js_function(
+            options_in_json)
+        return options_with_js_functions
+    else:
+        return options_in_json
