+++ pyecharts/utils.py
@@ -2,11 +2,7 @@
 from __future__ import unicode_literals
 
 import codecs
-import datetime
 import os
-import json
-
-import pyecharts.javascript as javascript
 
 
 def get_resource_dir(*paths):
@@ -29,40 +25,6 @@ def write_utf8_html_file(file_name, html_content):
     """
     with codecs.open(file_name, 'w+', encoding='utf-8') as f:
         f.write(html_content)
-
-
-class UnknownTypeEncoder(json.JSONEncoder):
-    """
-    UnknownTypeEncoder`类用于处理数据的编码，使其能够被正常的序列化
-    """
-
-    def default(self, obj):
-        if isinstance(obj, (datetime.datetime, datetime.date)):
-            return obj.isoformat()
-        else:
-            # Pandas and Numpy lists
-            try:
-                return obj.astype(float).tolist()
-            except Exception:
-                try:
-                    return obj.astype(str).tolist()
-                except Exception:
-                    return json.JSONEncoder.default(self, obj)
-
-
-def json_dumps(data, indent=0):
-    """ json 序列化编码处理
-
-    :param data: 字典数据
-    :param indent: 缩进量
-    """
-    options_in_json = json.dumps(data, indent=indent, cls=UnknownTypeEncoder)
-    if javascript.has_functions():
-        options_with_js_functions = javascript.unescape_js_function(
-            options_in_json)
-        return options_with_js_functions
-    else:
-        return options_in_json
 
 
 def to_css_length(x):
