+++ test/test_utils.py
@@ -1,18 +1,14 @@
 # coding=utf-8
 from __future__ import unicode_literals
 
-import codecs
-import json
 import os
-from datetime import date
+import codecs
 
-import numpy as np
 from nose.tools import eq_
 
 from pyecharts.utils import (
     write_utf8_html_file,
     get_resource_dir,
-    json_dumps,
     merge_js_dependencies
 )
 
@@ -30,20 +26,6 @@ def test_write_utf8_html_file():
     with codecs.open(file_name, 'r', 'utf-8') as f:
         actual_content = f.read()
         eq_(content, actual_content)
-
-
-def test_json_encoder():
-    """
-    Test json encoder.
-    :return:
-    """
-    data = date(2017, 1, 1)
-    eq_(json.dumps({'date': '2017-01-01', 'a': '1'}, indent=0),
-        json_dumps({'date': data, 'a': '1'}))
-
-    data2 = {'np_list': np.array(['a', 'b', 'c'])}
-    data2_e = {'np_list': ['a', 'b', 'c']}
-    eq_(json.dumps(data2_e, indent=0), json_dumps(data2))
 
 
 class MockChart(object):
